backend/requirements.txt
langchain==0.3.25
langchain_huggingface==0.2.0
langchain_openai==0.3.16
pandas==2.2.3
pymongo==4.12.1
Requests==2.32.3
